ci: add GitHub Actions workflow (build, fmt, clippy)
- Matrix: ubuntu-latest + macos-latest
- Rust nightly toolchain (edition 2024)
- cargo fmt --check (formatting)
- cargo clippy -D warnings (linting)
- Swatinem/rust-cache for dependency caching

+++ crates/webpilot-cli/src/commands/frames.rs
@@ -57,7 +57,7 @@ async fn list_frames(output_mode: OutputMode) -> Result<()> {
         } => match output_mode {
             OutputMode::Human => {
                 for f in &frames {
-                    let marker = if f.frame_id as i64 == active_frame_id {
+                    let marker = if f.frame_id == active_frame_id {
                         "*"
                     } else {
                         " "
